+++ offscreen.js
@@ -107,5 +107,3 @@ chrome.runtime.onMessage.addListener((message, _sender, sendResponse) => {
     return true;
   }
 });
-
-void initSdk();